backend/requirements.txt
# ==============================================
# バズ動画リサーチくん - バックエンド依存パッケージ
# Python 3.9+ 互換
# ==============================================

# Web Framework
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0

# Validation & Settings
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# HTTP Client (YouTube API)
httpx>=0.24.0,<1.0.0

# Claude API (バズ要因分析)
anthropic>=0.20.0,<1.0.0

# Retry Logic
tenacity>=8.0.0,<10.0.0

# Rate Limiting
slowapi>=0.1.9,<1.0.0

# Caching
cachetools>=5.0.0,<6.0.0

# Structured Logging
python-json-logger>=2.0.0,<4.0.0

# Metrics
prometheus-fastapi-instrumentator>=6.0.0,<8.0.0

# Supabase (認証・データベース)
supabase>=2.0.0,<3.0.0
PyJWT>=2.0.0,<3.0.0

# Development
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
